Restructure front-end assets for Webpack
It's time to move all the front-end assets to where Webpacker expects
them to be. I've also made some other minor changes as part of this
commit. Still more to do to completely remove the asset pipeline.
