+++ app/javascript/packs/application.js
@@ -12,7 +12,12 @@
 // them with the image_pack_tag helper in views (e.g <%= image_pack_tag 'rails.png' %>)
 // or the `imagePath` JavaScript helper below.
 //
-// const images = require.context('../images', true)
-// const imagePath = (name) => images(name, true)
+const images = require.context('../images', true)
+const imagePath = (name) => images(name, true)
 
-import '../css/application.css'
+import 'core-js/stable'
+import 'regenerator-runtime/runtime'
+import '../src/rails.js'
+import '../src/application.js'
+import '../src/application.css'
+import '../src/coderay.css'
