+++ app/javascript/src/rails.js
