+++ app/views/home/index.html.erb
@@ -1,6 +1,6 @@
 <h1>Home</h1>
 <div class="flex items-center px-6 py-4">
-  <%= image_tag("KeithDec2017.jpg", :alt => "photo", class: "block mx-auto sm:mx-0 sm:flex-shrink-0 h-auto sm:h-24 rounded-full") %>
+  <%= image_pack_tag("keith.jpg", :alt => "photo", class: "block mx-auto sm:mx-0 sm:flex-shrink-0 h-auto sm:h-24 rounded-full") %>
 </div>
 <h2>Who am I?</h2>
 <p>Glad you asked.</p>
