+++ app/views/layouts/admin_layout.html.erb
@@ -12,14 +12,14 @@
       <script src="//cdnjs.cloudflare.com/ajax/libs/html5shiv/3.6.1/html5shiv.js" type="text/javascript"></script>
     <![endif]-->
 
-    <%= stylesheet_pack_tag "application", "data-turbolinks-track": "reload" %>
-    <%= javascript_pack_tag "application", "data-turbolinks-track": "reload" %>
+    <%= stylesheet_pack_tag "application" %>
+    <%= javascript_pack_tag "application" %>
 
     <!-- Note: code removed for other icons until they are available -->
 
     <!-- For all other devices -->
     <!-- Size should be 32 x 32 pixels -->
-    <%= favicon_link_tag 'favicon.ico', :rel => 'shortcut icon' %>
+    <%= favicon_pack_tag 'favicon.ico', :rel => 'shortcut icon' %>
   </head>
   <body>
     <div class="w-screen bg-blue-900">
@@ -56,11 +56,6 @@
     <div class="container mx-auto px-12 pb-6">
       <p>&copy; <%= Time.zone.now.year.to_s %> Keith Pitty, all rights reserved.</p>
     </div>
-
-    <!-- Javascripts
-    ================================================== -->
-    <!-- Placed at the end of the document so the pages load faster -->
-    <%= javascript_include_tag "application" %>
 
   </body>
 </html>
