+++ config/webpack/environment.js
@@ -1,3 +1,9 @@
 const { environment } = require('@rails/webpacker')
 
+const webpack = require('webpack');
+environment.plugins.append('Provide', new webpack.ProvidePlugin({
+  $: 'jquery',
+  jQuery: 'jquery'
+}))
+
 module.exports = environment
