+++ postcss.config.js
@@ -8,7 +8,7 @@ module.exports = {
   plugins: [
     require('postcss-import'),
     require('postcss-flexbugs-fixes'),
-    require('tailwindcss')('./app/javascript/css/tailwind.js'),
+    require('tailwindcss')('./app/javascript/src/tailwind.js'),
     require('autoprefixer'),
     require('postcss-preset-env')({
       autoprefixer: {
@@ -16,8 +16,6 @@ module.exports = {
       },
       stage: 3
     }),
-    ...process.env.NODE_ENV === 'production'
-      ? [purgecss]
-      : []
+    ...process.env.NODE_ENV === 'production' ? [purgecss] : []
   ]
 }
